# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bengkel Sampah API",
        "description": "Dokumentasi API untuk sistem OTP dan autentikasi Bengkel Sampah",
        "contact": {
            "name": "Tim Bengkel Sampah",
            "email": "support@bengkelsampah.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Server utama"
        }
    ],
    "paths": {
...
        "/api/detail-profile": {
            "get": {
                "tags": [
                    "Profil"
                ],
                "summary": "Dapatkan data profil detail pengguna",
                "description": "Dapatkan nama pengguna dan daftar alamat",
                "operationId": "getDetailProfileData",
                "responses": {
                    "200": {
                        "description": "Operasi berhasil",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "data": {
                                            "properties": {
                                                "nama": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "addresses": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "nama": {
                                                                "type": "string",
                                                                "example": "John Doe"
                                                            },
                                                            "nomor_handphone": {
                                                                "type": "string",
                                                                "example": "081234567890"
                                                            },
                                                            "label_alamat": {
                                                                "type": "string",
                                                                "example": "Rumah"
                                                            },
                                                            "provinsi": {
                                                                "type": "string",
                                                                "example": "DKI Jakarta"
                                                            },
                                                            "kota_kabupaten": {
                                                                "type": "string",
                                                                "example": "Jakarta Selatan"
                                                            },
                                                            "kecamatan": {
                                                                "type": "string",
                                                                "example": "Kebayoran Baru"
                                                            },
                                                            "kode_pos": {
                                                                "type": "string",
                                                                "example": "12120"
                                                            },
                                                            "detail_lain": {
                                                                "type": "string",
                                                                "example": "Jl. Kebayoran Baru No. 123"
                                                            },
                                                            "is_default": {
                                                                "type": "boolean",
                                                                "example": true
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak terautentikasi. Token tidak diberikan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pengguna tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/edit-profile": {
            "put": {
                "tags": [
                    "Profil"
                ],
                "summary": "Perbarui profil pengguna",
                "description": "Perbarui nama, identifier, dan kata sandi pengguna. Jika memperbarui identifier, verifikasi OTP diperlukan.",
                "operationId": "updateProfile",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "identifier": {
                                        "type": "string",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "otp": {
                                        "description": "Diperlukan jika memperbarui identifier",
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profil berhasil diperbarui",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Profil berhasil diperbarui"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Tidak terautentikasi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tidak terautentikasi. Token tidak diberikan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Pengguna tidak ditemukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Pengguna tidak ditemukan."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data yang diberikan tidak valid."
                                        },
                                        "errors": {
                                            "properties": {
                                                "name": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "identifier": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "password": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "otp": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Kode OTP tidak valid",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Kode OTP tidak valid."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}